refactor: split oversized test files into per-module test files
Split 3 large test files (1107, 944, 824 LOC) into smaller focused
files, each under 400 LOC:

hierarchical_memory/tests.rs → tests/{mod,storage,search,statistics,
  consolidation,pattern}_tests.rs

cognitive_memory/tests.rs → tests/{mod,sensory,working,episodic,
  semantic,procedural,prospective,integration}_tests.rs

security/tests.rs → tests/{mod,query_validator,credential_scrubber,
  scope_enforcer,integration}_tests.rs

All 490 tests continue to pass. No logic changes.
