+++ rust/amplihack-memory/src/cognitive_memory/mod.rs
@@ -126,5 +126,4 @@ impl CognitiveMemory {
 // ---------------------------------------------------------------------------
 
 #[cfg(test)]
-#[path = "tests.rs"]
 mod tests;
