+++ rust/amplihack-memory/src/cognitive_memory/tests/mod.rs
@@ -1,0 +1,7 @@
+mod episodic_tests;
+mod integration_tests;
+mod procedural_tests;
+mod prospective_tests;
+mod semantic_tests;
+mod sensory_tests;
+mod working_tests;
