+++ rust/amplihack-memory/src/hierarchical_memory/tests/mod.rs
@@ -1,0 +1,5 @@
+mod consolidation_tests;
+mod pattern_tests;
+mod search_tests;
+mod statistics_tests;
+mod storage_tests;
