+++ rust/amplihack-memory/src/security/tests/mod.rs
@@ -1,0 +1,4 @@
+mod credential_scrubber_tests;
+mod integration_tests;
+mod query_validator_tests;
+mod scope_enforcer_tests;
